Use Khronos' OpenXR loader from Maven when supported
Most of the platforms used to distribute their own
implementations/builds of the OpenXR loader. It was traditionally
part of the propietary SDKs distributed by vendors. As they were
under an EULA we had to keep them in a private repository only
available to core devs (obviously any other external dev could
download them on their own).

More recently, and driven by the AOSP flavor effort, we
started to build the Khronos OpenXR loader from sources.
Fortunatelly we got a well documented report explaining how
to use the loader directly from the central Maven repository.
This greatly simplifies the build process and also improves
the open-source feel of the project by reducing our deps with
the third-party repo.

So far this option is now available for the following flavors:
* Meta Quest2, Quest3, QuestPro (requires firmware v62+)
* Lynx R1
* Magic Leap 2

It does not work for neither HVR, nor SnapdragonSpaces based
devices nor Pico. For these ones we still need to rely on the
loader from the SDK.

Fixes #1394